Complete functionality to load all locations
